CHG: Project settings have been changed to use "MinGW" build system. Still some issues I need to resolve here CHG: Manager now has virtual AddObjects() method because I needed 2 for estimates::Manager() CHG: doBuild.bat now returns a value (doBuild.sh should.. but it doesn't) so it can be used for a build system CHG: Updated how to build iSAM.txt with new instruction ADD: base::Object now allows the register and query of estimables ADD: configuration::Loader now supports loading @estimate blocks ADD: estimates::Info, isam::Estimates have both been implemented ADD: Model now validates and builds estimates::Manager ADD: Most processes now register estimables ADD: estimate_summary report ADD: All selectivities now register their estimables ADD: Test Case has been updated to load an estimable and print the report ADD: utilities::ToInline<>() method to allow inline conversions of types

+++ source/Processes/Children/RecruitmentConstant.h
@@ -44,7 +44,7 @@ private:
   // Members
   vector<string>              category_names_;
   map<string, Double>         proportions_;
-  unsigned                    r0_;
+  Double                      r0_;
   unsigned                    age_;
   CategoriesWithAgePtr        partition_;
 };
